+++ src/demonstrations/excludingdemographic.py
@@ -1,4 +1,3 @@
-
 from typing import List
 
 from tqdm import tqdm
@@ -8,7 +7,7 @@ import pandas as pd
 from .demographicdemonstration import DemographicDemonstration
 
 
-class WithinDemographic(DemographicDemonstration):
+class ExcludingDemographic(DemographicDemonstration):
     def __init__(self, shots: int = 16) -> None:
         super().__init__(shots)
 
@@ -18,12 +17,13 @@ class WithinDemographic(DemographicDemonstration):
         test_df: pd.DataFrame,
         overall_demographics: List[str],
     ) -> List[str]:
+        
         set_of_overall_demographics = set(overall_demographics)
 
         train_df["filtered_demographics"] = train_df["demographics"].apply(
             lambda x: self.filter_demographics(x, set_of_overall_demographics)
         )
-        test_df["filtered_demographics"] = train_df["demographics"].apply(
+        test_df["filtered_demographics"] = test_df["demographics"].apply(
             lambda x: self.filter_demographics(x, set_of_overall_demographics)
         )
 
@@ -33,13 +33,14 @@ class WithinDemographic(DemographicDemonstration):
 
         demonstrations = []
 
-        pre_computed_inclusions = dict()
+        pre_computed_exclusions = dict()
 
         for demographic in set_of_overall_demographics:
-            pre_computed_inclusions[demographic] = train_df[train_df.filtered_demographics == demographic]
+            pre_computed_exclusions[demographic] = train_df[~(train_df.filtered_demographics == demographic)]
+        
 
         for row in tqdm(test_df.itertuples()):
-            filtered_df = pre_computed_inclusions[row.filtered_demographics]
+            filtered_df = pre_computed_exclusions[row.filtered_demographics]
 
             train_dems = filtered_df["prompts"].sample(n=self.shots).tolist()
