+++ src/demonstrations/stratifiedsampler.py
@@ -1,17 +1,26 @@
-import random
+import math
 
-from typing import List
-
-from tqdm import tqdm
+from typing import List, Set
 
 import pandas as pd
+
+from tqdm import tqdm
 
 from .demographicdemonstration import DemographicDemonstration
 
 
-class ExcludingDemographic(DemographicDemonstration):
+class StratifiedSampler(DemographicDemonstration):
     def __init__(self, shots: int = 16) -> None:
         super().__init__(shots)
+
+    def stratified_sample_df(
+        self, df: pd.DataFrame, col: str, n_samples: int, number_of_demographics: int
+    ):
+        df_ = df.groupby(col).apply(
+            lambda x: x.sample(math.ceil(n_samples / number_of_demographics))
+        )
+        df_.index = df_.index.droplevel(0)
+        return df_
 
     def create_demonstrations(
         self,
@@ -19,13 +28,13 @@ class ExcludingDemographic(DemographicDemonstration):
         test_df: pd.DataFrame,
         overall_demographics: List[str],
     ) -> List[str]:
-        
+
         set_of_overall_demographics = set(overall_demographics)
 
         train_df["filtered_demographics"] = train_df["demographics"].apply(
             lambda x: self.filter_demographics(x, set_of_overall_demographics)
         )
-        test_df["filtered_demographics"] = train_df["demographics"].apply(
+        test_df["filtered_demographics"] = test_df["demographics"].apply(
             lambda x: self.filter_demographics(x, set_of_overall_demographics)
         )
 
@@ -35,15 +44,12 @@ class ExcludingDemographic(DemographicDemonstration):
 
         demonstrations = []
 
-        pre_computed_exclusions = dict()
-
-        for demographic in set_of_overall_demographics:
-            pre_computed_exclusions[demographic] = train_df[~(train_df.filtered_demographics == demographic)]
-
         for row in tqdm(test_df.itertuples()):
-            filtered_df = pre_computed_exclusions[row.filtered_demographics]
+            train_dems = self.stratified_sample_df(
+                train_df, "filtered_demographics", self.shots, len(set_of_overall_demographics)
+            )
 
-            train_dems = filtered_df["prompts"].sample(n=self.shots).tolist()
+            train_dems = train_dems["prompts"].tolist()[: self.shots]
 
             demonstrations.append("\n\n".join(train_dems) + "\n\n" + row.prompts)
