+++ src/demonstrations/withindemographic.py
@@ -1,26 +1,16 @@
-import math
 
-from typing import List, Set
-
-import pandas as pd
+from typing import List
 
 from tqdm import tqdm
+
+import pandas as pd
 
 from .demographicdemonstration import DemographicDemonstration
 
 
-class StratifiedSampler(DemographicDemonstration):
+class WithinDemographic(DemographicDemonstration):
     def __init__(self, shots: int = 16) -> None:
         super().__init__(shots)
-
-    def stratified_sample_df(
-        self, df: pd.DataFrame, col: str, n_samples: int, number_of_demographics: int
-    ):
-        df_ = df.groupby(col).apply(
-            lambda x: x.sample(math.ceil(n_samples / number_of_demographics))
-        )
-        df_.index = df_.index.droplevel(0)
-        return df_
 
     def create_demonstrations(
         self,
@@ -28,13 +18,12 @@ class StratifiedSampler(DemographicDemonstration):
         test_df: pd.DataFrame,
         overall_demographics: List[str],
     ) -> List[str]:
-
         set_of_overall_demographics = set(overall_demographics)
 
         train_df["filtered_demographics"] = train_df["demographics"].apply(
             lambda x: self.filter_demographics(x, set_of_overall_demographics)
         )
-        test_df["filtered_demographics"] = train_df["demographics"].apply(
+        test_df["filtered_demographics"] = test_df["demographics"].apply(
             lambda x: self.filter_demographics(x, set_of_overall_demographics)
         )
 
@@ -44,12 +33,15 @@ class StratifiedSampler(DemographicDemonstration):
 
         demonstrations = []
 
-        for row in tqdm(test_df.itertuples()):
-            train_dems = self.stratified_sample_df(
-                train_df, "filtered_demographics", self.shots, len(set_of_overall_demographics)
-            )
+        pre_computed_inclusions = dict()
 
-            train_dems = train_dems["prompts"].tolist()[: self.shots]
+        for demographic in set_of_overall_demographics:
+            pre_computed_inclusions[demographic] = train_df[train_df.filtered_demographics == demographic]
+
+        for row in tqdm(test_df.itertuples()):
+            filtered_df = pre_computed_inclusions[row.filtered_demographics]
+
+            train_dems = filtered_df["prompts"].sample(n=self.shots).tolist()
 
             demonstrations.append("\n\n".join(train_dems) + "\n\n" + row.prompts)
